+++ src/main/java/abstractions/selector/Human.java
@@ -63,7 +63,7 @@ public class Human extends AbstractSelector {
     }
 
     @Override
-    public List<MutationInterface> select(final TreeMap<Integer, List<MutationInterface>> evaluatedMutations) {
+    public List<MutationInterface> select(final TreeMap<Double, List<MutationInterface>> evaluatedMutations) {
         return Lists.newArrayList(this.askForSelection(Lists.newArrayList(Iterables.concat(evaluatedMutations.values()))));
     }
