+++ src/main/java/abstractions/selector/LastItem.java
@@ -29,7 +29,7 @@ public class LastItem extends AbstractSelector {
     }
 
     @Override
-    public List<MutationInterface> select(final TreeMap<Integer, List<MutationInterface>> evaluatedMutations) {
+    public List<MutationInterface> select(final TreeMap<Double, List<MutationInterface>> evaluatedMutations) {
         return evaluatedMutations.lastEntry().getValue();
     }
